fix: remove fail story in favor of test
