+++ packages/visualizations-react/jest.config.js
@@ -4,6 +4,7 @@ module.exports = {
     testEnvironment: 'jsdom',
     moduleNameMapper: {
         src: '<rootDir>/src',
+        'stories/(.*)': '<rootDir>/stories/$1',
         reactify: '<rootDir>/src/reactify',
         // https://jestjs.io/docs/webpack#handling-static-assets
         '\\.css$': '<rootDir>/test/__mocks__/styleMock.js',
