+++ packages/visualizations-react/package.json
@@ -77,6 +77,7 @@
         "@storybook/react-vite": "^7.6.17",
         "@testing-library/jest-dom": "^5.11.9",
         "@testing-library/react": "^12.1.5",
+        "@testing-library/user-event": "^14.5.2",
         "@turf/turf": "^6.5.0",
         "@types/geojson": "^7946.0.10",
         "@types/jest": "^29.0.2",
